index data

+++ convex/parties.ts
@@ -2,9 +2,17 @@ import { v } from 'convex/values';
 import { mutation, query } from './_generated/server';
 
 export const get = query({
-  args: {},
   handler: async (ctx) => {
-    return await ctx.db.query('parties').collect();
+    const identity = await ctx.auth.getUserIdentity();
+
+    if (!identity) {
+      throw new Error('Failed to fetch user information');
+    }
+
+    return await ctx.db
+      .query('parties')
+      .withIndex('by_createdBy', (q) => q.eq('createdBy', identity.subject))
+      .collect();
   },
 });
 
@@ -19,15 +27,34 @@ export const post = mutation({
     ),
   },
   handler: async (ctx, args) => {
-    return await ctx.db.insert('parties', args);
+    const identity = await ctx.auth.getUserIdentity();
+
+    if (!identity) {
+      throw new Error('Failed to fetch user information');
+    }
+
+    return await ctx.db.insert('parties', {
+      ...args,
+      createdBy: identity.subject,
+      updatedAt: Date.now(),
+    });
   },
 });
 
 export const deleteParty = mutation({
   args: {
     id: v.id('parties'),
+    createdBy: v.string(),
   },
   handler: async (ctx, args) => {
+    const identity = await ctx.auth.getUserIdentity();
+
+    if (!identity) {
+      throw new Error('Failed to fetch user information');
+    }
+
+    if (identity.subject !== args.createdBy) return;
+
     return await ctx.db.delete(args.id);
   },
 });
@@ -42,11 +69,23 @@ export const patchParty = mutation({
         name: v.string(),
       })
     ),
+    createdBy: v.string(),
   },
   handler: async (ctx, args) => {
+    const identity = await ctx.auth.getUserIdentity();
+
+    if (!identity) {
+      throw new Error('Failed to fetch user information');
+    }
+
+    if (identity.subject !== args.createdBy) {
+      throw new Error('Prevented illegal patch');
+    }
+
     return await ctx.db.patch(args.id, {
       name: args.name,
       heroes: args.heroes,
+      updatedAt: Date.now(),
     });
   },
 });
